# backend/docs/swagger/swagger.yaml
basePath: /api
definitions:
  octotify_internal_handler_dto.AuthResp:
    description: 认证成功后返回的 Access Token、Refresh Token 和用户信息
    properties:
      access_token:
        description: 访问令牌，有效期 1 小时
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        description: 刷新令牌，有效期 7 天
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        allOf:
        - $ref: '#/definitions/octotify_internal_handler_dto.UserDTO'
        description: 用户信息
    type: object
...
  octotify_internal_handler_dto.LoginReq:
    description: 用户登录时使用的请求参数
    properties:
      password:
        description: 密码，8-128 个字符，需包含大小写字母和数字
        example: P@ssw0rd123
        type: string
      username:
        description: 用户名，3-64 个字符，仅允许字母、数字和下划线
        example: octotify
        type: string
    required:
    - password
    - username
    type: object
...
  octotify_internal_handler_dto.PushMessageReq:
    description: 外部系统通过 Source Token 推送消息时使用的请求参数
    properties:
      message:
        description: 消息内容，不能为空
        example: 'Build #123 passed'
        minLength: 1
        type: string
      title:
        description: 消息标题，1-256 个字符
        example: CI Build
        maxLength: 256
        minLength: 1
        type: string
    required:
    - message
    - title
    type: object
  octotify_internal_handler_dto.PushResponse:
    description: 消息推送接口的返回结果，包含各渠道的推送详情
    properties:
      failed:
        description: 推送失败数
        type: integer
      results:
        description: 各渠道推送结果详情
        items:
          $ref: '#/definitions/octotify_internal_handler_dto.PushResult'
        type: array
      success:
        description: 推送成功数
        type: integer
      total:
        description: 推送渠道总数
        type: integer
    type: object
  octotify_internal_handler_dto.PushResult:
    description: 并发推送到多个渠道时，单个渠道的推送结果
    properties:
      channel_id:
        description: 渠道 ID
        type: integer
      channel_name:
        description: 渠道名称
        type: string
      error:
        description: 失败时的错误信息
        type: string
      message_id:
        description: 消息 ID
        type: integer
      success:
        description: 是否推送成功
        type: boolean
    type: object
  octotify_internal_handler_dto.RefreshReq:
    description: 使用 Refresh Token 获取新的 Access Token
    properties:
      refresh_token:
        description: Refresh Token，登录时获取
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
...
  octotify_internal_handler_dto.UserDTO:
    description: 用户基本信息，用于登录/注册/刷新令牌等接口返回
    properties:
      created_at:
        description: 创建时间（Unix 时间戳）
        example: 1714636800
        type: integer
      id:
        description: 用户 ID
        example: 1
        type: integer
      username:
        description: 用户名
        example: octotify
        type: string
    type: object
...
  octotify_pkg_response.Response:
    description: 所有 API 接口返回的统一格式，包含业务状态码、消息和数据
    properties:
      code:
        description: 业务状态码，0 表示成功，非 0 表示失败
        type: integer
      data:
        description: 业务数据，成功时返回业务数据，失败时可能为空或错误详情
      msg:
        description: 提示信息，成功时为"请求成功"，失败时为错误描述
        type: string
    type: object
host: localhost:34123
info:
  contact:
    name: OctoTify API Support
    url: https://github.com/OctoTify
  description: |-
    OctoTify 是一个消息总线平台，支持多种消息来源和推送渠道。
    核心功能：消息来源管理、推送渠道管理、消息推送与记录。
    本文档为前端开发人员提供完整的 API 接口说明，包含请求参数、响应格式、错误码等详细信息。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OctoTify API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: |-
        使用用户名和密码登录系统，登录成功后返回 Access Token 和 Refresh Token。
        Access Token 有效期 1 小时，用于调用需要认证的接口。
        Refresh Token 有效期 7 天，用于刷新 Access Token。
        ## 密码要求
        - 8-128 个字符
        - 必须包含小写字母
        - 必须包含大写字母
        - 必须包含数字
        ## 使用场景
        1. 用户首次访问系统时进行登录
        2. Refresh Token 过期后重新登录
        ## 注意事项
        - 登录成功后请妥善保存 Access Token 和 Refresh Token
        - Access Token 过期后请使用 Refresh Token 调用刷新接口
        - 连续多次登录失败不会锁定账户
        ## 错误码说明
        - 110200: 用户名或密码错误
        - 110201: 登录失败
      parameters:
      - description: 登录请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回认证信息
          schema:
            allOf:
            - $ref: '#/definitions/octotify_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/octotify_internal_handler_dto.AuthResp'
              type: object
      summary: 用户登录
      tags:
      - 用户认证
  /auth/logout:
    post:
      consumes:
      - application/json
      description: |-
        撤销当前用户的所有 Refresh Token，退出登录后需要重新登录才能访问系统。
        ## 注意事项
        - Access Token 在过期前仍有效（JWT 无状态特性，最长 1 小时后自动失效）
        - 前端需主动清除本地存储的 Access Token 和 Refresh Token
        - 退出登录后，所有使用旧 Refresh Token 的刷新请求都会失败
        ## 错误码说明
        - 100001: 未提供认证令牌或认证令牌无效或已过期
        - 110304: 退出登录失败
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            $ref: '#/definitions/octotify_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 退出登录
      tags:
      - 用户认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: |-
        使用 Refresh Token 获取新的 Access Token 和 Refresh Token。
        每次刷新都会返回新的 Access Token 和 Refresh Token，旧 Refresh Token 继续有效。
        ## 使用场景
        1. Access Token 过期（1 小时后）时调用此接口
        2. 前端应在请求收到 401 响应时自动调用此接口
        ## 注意事项
        - Refresh Token 有效期 7 天，过期后需重新登录
        - 刷新成功后请更新本地存储的 Access Token 和 Refresh Token
        - 如果 Refresh Token 已被撤销（如退出登录、修改密码），则无法刷新
        ## 错误码说明
        - 110300: 刷新令牌无效
        - 110301: 刷新令牌已撤销
        - 110302: 刷新令牌已过期
        - 110303: 刷新令牌失败
      parameters:
      - description: 刷新令牌请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.RefreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/octotify_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/octotify_internal_handler_dto.AuthResp'
              type: object
      summary: 刷新令牌
      tags:
      - 用户认证
...
  /push/{token}:
    post:
      consumes:
      - application/json
      description: |-
        外部系统通过 Source Token 推送消息到平台，消息会被并发推送到所有绑定的有效渠道。
        ## 认证方式
        使用 Source Token 进行认证，而非用户 Access Token。
        Token 格式：src{uuid}，例如：src0196a3b2c4d50000a1b2c3d4e5f67890
        ## 推送流程
        1. 系统验证 Source Token 有效性
        2. 查找该来源绑定的所有启用状态的渠道
        3. 并发推送到所有渠道
        4. 返回各渠道的推送结果
        ## 使用场景
        1. CI/CD 流水线推送构建结果
        2. 监控系统推送告警信息
        3. 业务系统推送通知消息
        ## 注意事项
        - Source Token 通过 URL 路径参数传递
        - 消息会并发推送到所有绑定的有效渠道
        - 返回结果包含每个渠道的推送状态
        - 推送失败不影响其他渠道的推送
        - 来源被停用后无法推送消息
        ## 错误码说明
        - 120400: 来源 Token 无效
        - 120401: 来源已停用
        - 120402: 来源不存在
        - 120403: 没有可用的推送渠道
        - 120404: 推送消息失败
      parameters:
      - description: 来源 Token，格式为 src{uuid}
        in: path
        name: token
        required: true
        type: string
      - description: 推送消息请求参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/octotify_internal_handler_dto.PushMessageReq'
      produces:
      - application/json
      responses:
        "200":
          description: 推送成功
          schema:
            allOf:
            - $ref: '#/definitions/octotify_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/octotify_internal_handler_dto.PushResponse'
              type: object
      security:
      - SourceTokenAuth: []
      summary: 推送消息
      tags:
      - 消息推送
...
securityDefinitions:
  BearerAuth:
    description: 输入格式：Bearer {access_token}，例如：Bearer eyJhbGciOiJSUzI1NiIs...
    in: header
    name: Authorization
    type: apiKey
  SourceTokenAuth:
    description: 推送消息时使用 Source Token，输入格式：Bearer src{uuid}，例如：Bearer src0196a3b2c4d50000a1b2c3d4e5f67890
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...